Fix /treaty drop-cap + reader-mode typography on server pages
`readerMarkdownComponents` was exported from `ReferendumStepper.tsx`, which
is a `"use client"` module. When the server-rendered `/treaty/page.tsx`
imported the components and passed them to `<ReactMarkdown>`, the
components crossed the React Server Components boundary as client
references — and their JSX wrappers were silently dropped. The rendered
`<p>` elements came out as bare `<p>WHEREAS, …</p>` with no classes:
no drop-cap, no Libre Baskerville, no `text-lg sm:text-[1.35rem]`.

Extract `readerMarkdownComponents` to its own non-"use client" module so
both server and client surfaces get the same prestige typography. The
client `ReferendumStepper` re-exports the same value for backward compat
with `TreatyContent.tsx` and the existing test mock.

Verified: bare `<p>WHEREAS` -> `<p class="mb-8 text-left text-lg leading-9 …drop-cap [font-family:var(--v0-font-libre-baskerville)] sm:text-[1.35rem]">WHEREAS`.
